==> 02-data.Rmd <==
# Data sources

The dataset used in this project was obtained from UCI Machine Learning Repository(http://archive.ics.uci.edu/ml/datasets/Student+Performance). It was originally generated by Paulo Cortez and then donated to the University of California Irvine database.

## Overview
This dataset approaches student achievement in secondary education of two Portuguese schools. Two datasets are provided regarding the performance in two distinct subjects: Mathematics (mat) and Portuguese language (por).

## Data Dictionary
Attributes for both datasets(One for math grade and the other for Portuguese grade):

```{r, echo = FALSE}
Description = readr::read_csv(here::here("./data/clean/description.csv"))
knitr::kable(Description[1:33, ], caption = "Columns and Descriptions",
             row.names = F,font_size = 12)
```

Note: Grades are related with the course subject, Math or Portuguese.
