==> 05-results.Rmd <==
---
output: html_document
editor_options: 
chunk_output_type: console
---
# Results

```{r echo=FALSE, message=FALSE, warning=FALSE}
library(ggplot2)
library(dplyr)
library(tidyverse)
library(vcd)
library(HH)
library(here)
library(readr)
library(psych)
library(ggalluvial)
```

## Students & Schools Analysis

This section will focus on non-academic related variables, i.e. analyzing students' background and demographic information by schools. It is noted that the data used in this section takes all rows after merging since our goal is not to evaluate students' performance.

### Age

```{r age}
stu <- readr::read_csv(here::here("./data/clean/nonaca.csv")) %>% 
  mutate(sex = ifelse(sex == "F","Female", "Male"),
         address = ifelse(address == "U","Urban ", "Rural"))
ggplot(stu, aes(factor(age))) +
  geom_bar(color = "black", 
           fill = "lightblue") +
  facet_wrap(~school) +
  xlab("Age") +
  ylab("Frequency") +
  ggtitle("Bar chart for the age by schools") +
  theme_bw() +
  theme(plot.title = element_text(hjust = 0.5),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())
```

The bar chart plot displays students' age distribution in the two schools. One of obvious properties is that **GP** has much **more** students than **MS**. Precisely, **GP** has a total of `r sum(ifelse(stu$school=="GP",1,0))` students of which accounts for `r round(sum(ifelse(stu$school=="GP",1,0))/dim(stu)[1],4)*100`% of the dataset. Possibly due to the small number of students, age range is more narrow in **MS** (17-19) than that in **GP** (15-22) except for 21.

In order to compare the age distribution effectively, we draw the relative frequency plot below.

```{r}
stu %>% 
  count(school, age) %>% 
  ungroup() %>% 
  group_by(school) %>% 
  mutate(per = n / sum(n)) %>% 
  dplyr::select(-n) %>% 
  ungroup() %>% 
  add_row(school = c("GP","MS","MS","MS","MS"), 
          age = c(21, 15, 16, 22, 20), 
          per = c(0, 0, 0, 0, 0)) %>% 
  ggplot(aes(factor(age), per * 100, fill = school)) + 
    geom_bar(stat='identity', position = "dodge")  + 
    xlab("Age") +
    ylab("Relative frequency in %") +
    ggtitle("Bar chart for the age by schools") +
    theme_bw() +
    theme(plot.title = element_text(hjust = 0.5),
          panel.grid.major.x = element_blank(),
          panel.grid.minor.x = element_blank())
```

In the relative frequency plot, it is clear that students in **GP** are generally younger than that in **MS**. In addition, age distribution in **MS** is centered that age group 18 is an absolute majority while age distribution in **GP** is spread that no age group accounts for more than one-third.

### Family Education Level

In this section, we will analyze the students' family education level in the two schools. Since the dataset offers the education level of both parents, we will use the higher one to indicate the family education level of a student, i.e. `max(Medu, Fedu)`.

```{r}
HH::likert(Pedu~., 
           stu %>%  mutate(Pedu = pmax(Medu, Fedu)) %>%
             count(school, Pedu, name = "Freq") %>% 
             pivot_wider(names_from = school, 
                         values_from = Freq) %>% 
             mutate(GP = 100 * GP / sum(GP),
                    MS = 100 * MS / sum(MS)) %>% 
             arrange(desc(Pedu)),
#           positive.order = TRUE,
           main = "Family education level in two schools",
           xlab = "percent", ylab = "")
```

According to the plot above, we can find that students' family education level in **GP** is generally higher than that in **MS**.

### Sex & Address

In this section, we are going to look at the students' distribution in terms of their gender and home address type (Urban or Rural).

```{r}
HH::likert(grp~., 
           union(stu %>% 
                   count(school, sex) %>% 
                   pivot_wider(names_from = school, values_from = n) %>% 
                   rename(grp = sex) %>% 
                   mutate(GP = 100 * GP / sum(GP),
                          MS = 100 * MS / sum(MS)),
                 stu %>% 
                   count(school, address) %>% 
                   pivot_wider(names_from = school, values_from = n) %>% 
                   rename(grp = address) %>% 
                   mutate(GP = 100 * GP / sum(GP),
                          MS = 100 * MS / sum(MS))%>% 
             arrange(desc(grp))),
#           positive.order = TRUE,
           main = "% of gender and address in two schools",
           xlab = "percent", ylab = "")
```

Both **GP** and **MS** have a **half-half** gender distribution. However, when it comes to students' home address in two schools, only around 20% of students in **GP** are from rural areas while there is no significant difference in students' address type in **MS**. Considering the fact that **GP** has a large sample size, we can conclude that rural students are less likely to choose **GP**.

### Reason of Chosing School

In this section, we are going to conduct multivariate analysis in terms of the association among schools, reasons, and family education level.

#### Reason & School

According to the plot below, there is an association between `school` and `reason`. The **Pearson's Chi-squared** test rejects the hypothesis that two variables are not relevant with a p-value of 0.006, which is consistent with our observation.

Consequently, we can obtain that **GP** has a better reputation in terms of students' choices while courses offered by **MS** is more attractive.

```{r}
mosaic(reason ~ school, 
       stu %>% count(school, reason, name = "Freq") %>%
         mutate(reason=as.factor(reason)) %>% 
         mutate(reason = fct_relevel(reason, "other", after = Inf)), direction = c("v", "h"))
```

#### Reason ~ Address + School

The plot below indicates that rural students are more likely to choose **MS** over **GP** due to home and other reasons. A possible explanation would be **MS**'s schedule is flexible and friendly to rural students. 

On the other hand, urban students tend to choose **MS** more often than **GP** because of `course`, which is consistent with what we obtain in the previous section.

```{r}
mosaic(reason ~ address + school, 
       stu %>% count(school, reason, address, name = "Freq") %>%
         mutate(reason=as.factor(reason)) %>% 
         mutate(reason = fct_relevel(reason, "other", after = Inf)), 
       direction = c("v", "v", "h"))
```

## Factors Affecting Students' Academic Performance

In this section, we will analyze factors that affect students' academic performance. Given that we have more than 40 variables in our dataset and all of them are categorical variables after conversion, we decide to use Pearson's Chi-squared Test to help us narrow down our variable list to be analyzed later. Here we are using the final (the third period) letter grade (`Math_Letter_Grade_G3` and `Por_Letter_Grade_G3` converted from `G3` based on the grading system in Portugal) for both **Math** and **Portuguese** to reflect students' academic performance. 

For each variable in our dataset, we conduct a Pearson's Chi-squared Test to see if there is an association between it and `Math_Letter_Grade_G3` or `Portuguese_Letter_Grade_G3`. We set \alpha with 0.01, and the following graphs show factors that have an impact on students' **Math** and **Portuguese** grades: 

```{r}
df = read.csv("./data/clean/df.csv")
test = function(x){
  l = c()
  for(i in 1:(ncol(x)-1)){
    if(chisq.test(x[,i],x[,ncol(x)])$p.value < 0.01){
      l = append(l,i)
    }
  }
  return(l)
}
Math_cor = df[,c(1:30,57)]
colnames(Math_cor) = str_remove(colnames(Math_cor),'\\.x')
Por_cor = df[,c(1:15,34:48,54)]
colnames(Por_cor) = str_remove_all(colnames(Por_cor),'\\.y')
list_math = test(Math_cor)
list_por = test(Por_cor)
MathCorMat = data.frame(factors = colnames(Math_cor)[list_math], cor = 1)
for(i in 1:length(list_math)){
  if(typeof(Math_cor[,list_math[i]]) == 'integer' | typeof(Math_cor[,list_math[i]]) == 'double'){
    MathCorMat[i,2] = cor(df$Math_G3,Math_cor[,list_math[i]])
  }else{
    MathCorMat[i,2] = cor(df$Math_G3,case_when(Math_cor[,list_math[i]] == 'yes' ~ 1, Math_cor[,list_math[i]] == 'no' ~ 0))
  }
}
MathCorMat['NP'] = case_when(MathCorMat$cor > 0 ~ 'Positive', MathCorMat$cor < 0 ~ 'Negative')
f1 = ggplot(MathCorMat, aes(y = cor, x = fct_reorder(factors,cor,.desc=TRUE), fill = NP)) + geom_bar(stat="identity") + xlab('factors') + ggtitle('Correlations between different factors and Math grade in the third period') + theme_bw() + theme(legend.position = 'none',plot.title = element_text(hjust = 0.5),panel.grid.major.x = element_blank(),panel.grid.minor.x = element_blank())

PorCorMat = data.frame(factors = colnames(Por_cor)[list_por], cor = 1)
for(i in 1:length(list_por)){
  if(typeof(Por_cor[,list_por[i]]) == 'integer' | typeof(Por_cor[,list_por[i]]) == 'double'){
    PorCorMat[i,2] = cor(df$Portuguese_G3,Por_cor[,list_por[i]])
  }else{
    PorCorMat[i,2] = cor(df$Portuguese_G3,case_when(Por_cor[,list_por[i]] == 'yes' ~ 1, Por_cor[,list_por[i]] == 'no' ~ 0))
  }
}
PorCorMat['NP'] = case_when(PorCorMat$cor > 0 ~ 'Positive', PorCorMat$cor < 0 ~ 'Negative')
f2 = ggplot(PorCorMat, aes(y = cor, x = fct_reorder(factors,cor,.desc=TRUE), fill = NP)) + geom_bar(stat="identity") + xlab('factors') + ggtitle('Correlations between different factors and Portuguese grade in the third period') + theme_bw() + theme(legend.position = 'none',plot.title = element_text(hjust = 0.5),panel.grid.major.x = element_blank(),panel.grid.minor.x = element_blank())

grid.arrange(f1,f2,nrow=2)
```

There are some common factors: `Medu` has a positive impact on both Math and Portuguese grade, while `schoolsup` and `failures` have a negative impact on the two grades. 

Now we can dig deeper into these variables one by one. 

### Variables Related to Math Grade

```{r}
p1 = grid.grabExpr(mosaic(Medu ~ Math_Letter_Grade_G3, Math_cor, direction = c("v", "h")))
p2 = grid.grabExpr(mosaic(failures ~ Math_Letter_Grade_G3, Math_cor, direction = c("v", "h")))
p3 = grid.grabExpr(mosaic(schoolsup ~ Math_Letter_Grade_G3, Math_cor, direction = c("v", "h")))
p4 = grid.grabExpr(mosaic(absences ~ Math_Letter_Grade_G3, Math_cor, direction = c("v", "h")))
grid.arrange(p1,p2,ncol=2)
```
Here are some observations from the plot above:

   * For `Medu`: there is an association between a student's mother’s education level and student's math letter grade. When the education for a student's mother is in level 3 or 4, the student is more likely to get a better grade(A or B). The is saying that a student's mother with high-level education can be helpful to the student's achievement in math.
   * For `failures`: students who got a grade of A or B never failed any courses in the past, and few of them have a small number of failures (less than 1). However, the proportion of students who failed more than 3 courses in the past increases rapidly in the class of grade C or D. This is reasonable that a student with many failed courses is prone to get a low grade in the future class.

```{r}
grid.arrange(p3,p4,ncol=2)
```
Observations:

   * For `schoolsup`: it is clear that the proportion of students who get extra educational support increases as `Math_Letter_Grade_G3` goes from A to D. This seems reasonable because students who got low grades need extra educational support for improvement.
   * For `absences`: the plot shows something unexpected. If we only focus on grade B, C and D, it indicates a trend that fewer absences from school lead to a higher grade in math class. But when we move to grade A, a reverse trend applies. We can see that students with grade A are more likely to be absent from school than students with grade B.

### Variables Related to Portuguese Grade
```{r,warning=FALSE}
p5 = grid.grabExpr(mosaic(Medu ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p6 = grid.grabExpr(mosaic(studytime ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p7 = grid.grabExpr(mosaic(failures ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p8 = grid.grabExpr(mosaic(schoolsup ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p9 = grid.grabExpr(mosaic(higher ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p10 = grid.grabExpr(mosaic(freetime ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p11 = grid.grabExpr(mosaic(goout ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p12 = grid.grabExpr(mosaic(Dalc ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
p13 = grid.grabExpr(mosaic(health ~ Por_Letter_Grade_G3, Por_cor,direction = c("v","h")))
grid.arrange(p5,p6, ncol = 2)
```
Observations:

   * For `Medu`: from the mosaic plot, we can find that the majority of students got grade in class B, C and D. If we focus on these classes only, the proportion of students whose mother's education is in level 3 or 4 decreases as we move from class B to class D(so as the proportion of `Medu` = 0 ,1, 2 increases). On the other hand, class A does not seem to follow this trend. However, given that the number of students who got grade A is only 11 and we have 333 students in total, we may need more data to analyze the hidden reason. But it is clear that the proportion of `Medu` = 3 or 4(high-level education) in class A is relatively high compared to class D and F. In conclusion, we expect to see a student get a higher grade if his/her mother has received higher-level education. \
   * For `studytime`: the plot reveals that students with higher grades used to spend more time on study. This is intuitively correct as students spend more time on study, which will increase their knowledge and self-assurance to score a higher grade. \

```{r}
grid.arrange(p7,p8,ncol=2)
```
Observations:

   * For `failures`: there is the same trend as how `failures` affect students' performance in math. We don't see much difference in class A and class B. But when we move to class C, D and F, the number of students who failed more than 2 classes in the past increases dramatically. Hence, here is the same idea that a student with failures in the past class is prone to get a lower grade in the future class. 
   * For `schoolsup`: same as the relationship between `schoolsup` and `Math_Letter_Grade_G3`, the proportion of students who get extra educational support increases when `Por_Letter_Grade_G3` goes from A to F. This seems reasonable because students who got low grades need extra educational support to help them improve.

```{r}
grid.arrange(p9,p10,ncol=2)
```
Observations:

   * For `higher`: the plot shows that students who do not want to pursue a higher level education finally got grade C or D. We cannot say much about a student who wants to take higher education, because he/she is equally likely to get grade A, B or C given that the proportion of `higher = yes` is almost the same in grade A, B and C. However, if a student does not want to take higher education, our data supports that he/she is likely to get a grade C or D. Grade F may be an exception here and we cannot conclude anything for it given that there is only one student get grade F. 
   * For `freetime`: there is no clear trend to show how `freetime` affects students' grades of Portuguese language. Class B and Class C look almost the same. An interesting fact is that, there is a boost in students' free time after school when we move to A, D or F, which indicates that a big increment in `freetime` may reflect two totally different situations: either a student did well in Portuguese exam or a student got low grade(D or F) in the exam. A possible explanation is that, generally, students with grade A study more strategically which increases their study efficiency and save more time on finishing their homework, so have more free time compared to students with grade B and C who spend more time on study with low efficiency so have less free time. And for students with grade D and F, they just spend less time on study so have more free time. 

```{r}
grid.arrange(p11,p12,ncol=2)
```

Observations:

   * For `goout`: there is a clear trend that students who often go out with friends(`goout = 4 or 5`) are less likely to get a high grade in Portuguese language exams. 
   * For `Dalc`: the mosaic plot shows that the proportion of `Dalc = 3, 4 or 5` increases in grade B, C, D and F which indicates that increment in alcohol consumption in the workday will have a negative impact on students' performance on Portuguese. 
```{r}
grid.arrange(p13)
```

Observations:

   * For `health`: there is a strong relationship between `health` and `Por_Letter_Grade`. Students with a higher grade may tend to be more stressed regarding their studies. Moreover, as we discovered before, students with a higher grade have longer weekly study time. Hence, unsurprisingly, we expect to see that students with a higher grade have worse health condition than those with a lower grade, as spending more time on study may have a negative impact on health.

## Grade in Different Periods

In this section, we are interested in the grades of Math and Portuguese in different periods. We tend to believe that there should be strong relationships among the exam grades in different periods, that is if a student can do well in G1 or G2, he/she is expected to do well in G3 as well. 

Let's take a look at the alluvial diagram below:

```{r}
df_new = read.csv("./data/clean/df.csv")
math_letter = df_new[,c(1,55:57)]
colnames(math_letter) = str_remove(colnames(math_letter),'Math\\_')
math_letter['subject'] = 'Math'
por_letter = df_new[,c(1,52:54)]
colnames(por_letter) = str_remove(colnames(por_letter),'Por\\_')
por_letter['subject'] = 'Portuguese'
combine = rbind(math_letter,por_letter)

ggplot(combine,aes(axis1 = school, axis2 = Letter_Grade_G1, axis3 = Letter_Grade_G2, axis4 = Letter_Grade_G3)) +
  geom_alluvium(aes(fill=subject))+ 
  geom_stratum(width = 1/12, alpha=.5) +
  geom_label(stat = "stratum", 
             aes(label = after_stat(stratum))) +
  scale_x_discrete(limits = c("school", "G1_Letter_Grade", "G2_Letter_Grade", "G3_Letter_Grade"),expand = c(.05, .05)) +
  scale_fill_brewer(type = "qual", palette = "Set1") +
  ggtitle("Alluvial diagram of students' grades through academic year") + 
  theme_bw() + 
  theme(legend.position = 'none',
        plot.title = element_text(hjust = 0.5),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())
```

The alluvial plot represents the change of students' grades over the academic year (from the first period to the final period). Most students get the same letter grade for G1, G2 and G3 which implies that if a student performs well in an exam, he/she is likely to get a high grade in the next exam. In this case, we believe that students' grades in different periods are highly correlated. We can verify this by computing the correlations. This time, we will use `Math_G1`, `Math_G2`, `Math_G3`, `Por_G1`, `Por_G2` and `Por_G3` which are numeric grades instead of letter ones.
```{r}
pairs.panels(df[,c(31:33,49:51)],
             method = "pearson",
             hist.col = "lightblue",
             density = TRUE, 
             ellipses = TRUE)
```

The pair plot reveals that there are strong correlations(approch 1) within math/Portuguese grades in different periods.
