==> 07-conclusion.Rmd <==
# Conclusion

## Students & School

Based on our analysis, we have the following conclusion about students' background and demographic information by schools:

   * Students in GP are generally younger than that in MS.
   * Students’ family education level in GP is generally higher than that in MS.
   * Rural students tend to choose MS over GP.
   * Based on students' choice, GP has a better reputation while courses offered by MS is more attractive.

## Academic Performance

After carefully exploring and visualizing our dataset, we conclude that:

   * `Medu`(mother's education) has a positive impact on students' math grade while `absences`(absence from school), `schoolsup`(extra educational support) and `failures`(number of failures in the past classes) have a negative impact on student's math grade.
   * `studytime` (weekly study time), `higher` (wants to take higher grade) and `Medu`(mother's education) have a positive impact on students' Portuguese language grade while `freetime`(free time after school), `goout`(frequency of going out with friends), `health`(current health status), `schoolsup`(extra educational support), `Dalc`(alcohol consumption in weekday) and `failures`(number of failures in the past classes) have a negative impact on students' Portuguese language grade. 

## Pattern in Students’ Grades

Our discovery suggests that there is a strong relationship among grades in different periods. G1, G2 and G3 are highly correlated.
